fix: GuideModal Configuration
