--- src/components/VoiceControlGuide.js
@@ -1,5 +1,0 @@
-const VoiceControlGuide = (guideImage, closeModal) => {
-  return <div>VoiceControlGuide</div>;
-};
-
-export default VoiceControlGuide;
