+++ src/pages/PageSingleRecipe.js
@@ -6,8 +6,7 @@ import YouTube from "react-youtube";
 import Notes from "./PageSingleRecipe.Notes";
 import Tips from "./PageSingleRecipe.Tips";
 import Note from "./PageSingleRecipe.MyNote";
-import ModalGuide from "../components/ModalGuide";
-import VoiceControlGuide from "../components/VoiceControlGuide";
+import ModalGuideSinglePage from "../components/ModalGuideSinglePage";
 import { useQuery } from "react-query";
 import { getRecipe } from "../api/recipeApi";
 import { videoOptions } from "../config/youtubeConfig";
@@ -15,7 +14,6 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faFileAudio } from "@fortawesome/free-regular-svg-icons";
 import useRecipeControlBySpeech from "../hooks/recipe-speech-control";
 import TypeWriter from "typewriter-effect";
-
 
 function PageSingleRecipe({ loginUserInfo, toggleTheme, theme }) {
   const [currentBoardPage, setBoardPage] = useState("notes");
@@ -90,11 +88,7 @@ function PageSingleRecipe({ loginUserInfo, toggleTheme, theme }) {
 
   return (
     <Container>
-      {isModalOpen && (
-        <ModalGuide>
-          <VoiceControlGuide />
-        </ModalGuide>
-      )}
+      {isModalOpen && <ModalGuideSinglePage />}
       <LeftSection>
         <NavigationPage>
           <StyledLinkButton
@@ -115,6 +109,10 @@ function PageSingleRecipe({ loginUserInfo, toggleTheme, theme }) {
             {theme === "light" && "Dark Mode ☾"}
             {theme === "dark" && "Light Mode ☀"}
           </ViewMode>
+          <StyledFontAwesomeIcon
+            icon={faFileAudio}
+            onClick={clickVoiceControlGuide}
+          />
           {isCommanding && (
             <RecodingBox>
               <RecordingStatus className="blob red"></RecordingStatus>
@@ -168,10 +166,7 @@ function PageSingleRecipe({ loginUserInfo, toggleTheme, theme }) {
                 꿀팁
               </Button>
             </ButtonLeft>
-            <StyledFontAwesomeIcon
-              icon={faFileAudio}
-              onClick={clickVoiceControlGuide}
-            />
+
             <ButtonRight>
               {currentBoardPage !== "myNote" && (
                 <Button
@@ -254,10 +249,8 @@ const VideoPlayer = styled.article`
 `;
 
 const StyledFontAwesomeIcon = styled(FontAwesomeIcon)`
-  position: absolute;
   font-size: 25px;
-  top: 3%;
-  right: 12%;
+  margin-right: 5px;
 
   cursor: pointer;
 `;
